Refactor logger and registry into separate files.
