+++ packages/logconsole/src/index.ts
@@ -1,5 +1,7 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
 
+export * from './logger';
+export * from './registry';
 export * from './tokens';
 export * from './widget';
