+++ packages/logconsole/src/widget.ts
@@ -1,317 +1,28 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
 
-import { StackedPanel, Widget, Panel } from '@phosphor/widgets';
+import { nbformat } from '@jupyterlab/coreutils';
 
-import { ISignal, Signal } from '@phosphor/signaling';
+import { OutputArea, IOutputPrompt } from '@jupyterlab/outputarea';
+
+import { IRenderMimeRegistry } from '@jupyterlab/rendermime';
 
 import { Kernel, KernelMessage } from '@jupyterlab/services';
 
-import { nbformat } from '@jupyterlab/coreutils';
-
-import {
-  OutputArea,
-  IOutputAreaModel,
-  OutputAreaModel,
-  IOutputPrompt
-} from '@jupyterlab/outputarea';
-
-import {
-  IRenderMimeRegistry,
-  IOutputModel,
-  OutputModel
-} from '@jupyterlab/rendermime';
-
 import { Message } from '@phosphor/messaging';
+
+import { ISignal, Signal } from '@phosphor/signaling';
+
+import { StackedPanel, Widget, Panel } from '@phosphor/widgets';
+
+import { LogOutputModel, LoggerOutputAreaModel } from './logger';
 
 import {
   ILogger,
   ILoggerChange,
-  ILogPayload,
   ILoggerRegistry,
   ILoggerRegistryChange
 } from './tokens';
-
-/**
- * Custom Notebook Output with timestamp member.
- */
-interface ITimestampedOutput extends nbformat.IBaseOutput {
-  /**
-   * Date & time when output is logged in integer representation.
-   */
-  timestamp: number;
-}
-
-/**
- * Custom Notebook Output with optional timestamp.
- */
-type IOutputWithTimestamp = nbformat.IOutput | ITimestampedOutput;
-
-/**
- * Log Output Model with timestamp which provides
- * item information for Output Area Model.
- */
-class LogOutputModel extends OutputModel {
-  /**
-   * Construct a LogOutputModel.
-   *
-   * @param options - The model initialization options.
-   */
-  constructor(options: LogOutputModel.IOptions) {
-    super(options);
-
-    this.timestamp = new Date(options.value.timestamp as number);
-  }
-
-  /**
-   * Date & time when output is logged.
-   */
-  timestamp: Date = null;
-}
-
-/**
- * Log Output Model namespace that defines initialization options.
- */
-namespace LogOutputModel {
-  export interface IOptions extends IOutputModel.IOptions {
-    value: IOutputWithTimestamp;
-  }
-}
-
-/**
- * Implementation of `IContentFactory` for Output Area Model
- * which creates LogOutputModel instances.
- */
-class LogConsoleModelContentFactory extends OutputAreaModel.ContentFactory {
-  /**
-   * Create a rendermime output model from notebook output.
-   */
-  createOutputModel(options: IOutputModel.IOptions): LogOutputModel {
-    return new LogOutputModel(options);
-  }
-}
-
-/**
- * A concrete implementation of ILogger.
- */
-export class Logger implements ILogger {
-  /**
-   * Construct a Logger.
-   *
-   * @param source - The name of the log source.
-   */
-  constructor(options: Logger.IOptions) {
-    this.source = options.source;
-    this.outputAreaModel = new LoggerOutputAreaModel({
-      contentFactory: new LogConsoleModelContentFactory(),
-      maxLength: options.maxLength
-    });
-  }
-
-  /**
-   * The maximum number of outputs stored.
-   *
-   * #### Notes
-   * Oldest entries will be trimmed to ensure the length is at most
-   * `.maxLength`.
-   */
-  get maxLength() {
-    return this.outputAreaModel.maxLength;
-  }
-  set maxLength(value: number) {
-    this.outputAreaModel.maxLength = value;
-  }
-
-  /**
-   * Number of outputs logged.
-   */
-  get length(): number {
-    return this.outputAreaModel.length;
-  }
-
-  /**
-   * A signal emitted when the list of log messages changes.
-   */
-  get logChanged(): ISignal<this, ILoggerChange> {
-    return this._logChanged;
-  }
-
-  /**
-   * A signal emitted when the rendermime changes.
-   */
-  get rendermimeChanged(): ISignal<this, void> {
-    return this._rendermimeChanged;
-  }
-
-  /**
-   * Log an output to logger.
-   *
-   * @param log - The output to be logged.
-   */
-  log(log: ILogPayload) {
-    const timestamp = new Date();
-    let output: nbformat.IOutput = null;
-
-    switch (log.type) {
-      case 'text':
-        output = {
-          output_type: 'display_data',
-          data: {
-            'text/plain': log.data
-          }
-        };
-        break;
-      case 'html':
-        output = {
-          output_type: 'display_data',
-          data: {
-            'text/html': log.data
-          }
-        };
-        break;
-      case 'output':
-        output = log.data;
-        break;
-      default:
-        break;
-    }
-
-    if (output) {
-      this.outputAreaModel.add({ ...output, timestamp: timestamp.valueOf() });
-      this._active = true;
-      this._logChanged.emit('append');
-    }
-  }
-
-  /**
-   * Clear all outputs logged.
-   */
-  clear() {
-    this.outputAreaModel.clear(false);
-    this._logChanged.emit('clear');
-  }
-
-  /**
-   * Rendermime to use when rendering outputs logged.
-   */
-  get rendermime(): IRenderMimeRegistry | null {
-    return this._rendermime;
-  }
-  set rendermime(value: IRenderMimeRegistry | null) {
-    if (value !== this._rendermime) {
-      this._rendermime = value;
-      this._rendermimeChanged.emit();
-    }
-  }
-
-  /**
-   * Whether the log has ever had a message.
-   */
-  get active(): boolean {
-    return this._active;
-  }
-
-  readonly source: string;
-  readonly outputAreaModel: LoggerOutputAreaModel;
-  private _logChanged = new Signal<this, ILoggerChange>(this);
-  private _rendermimeChanged = new Signal<this, void>(this);
-  private _rendermime: IRenderMimeRegistry | null = null;
-  private _active = false;
-}
-
-export namespace Logger {
-  export interface IOptions {
-    /**
-     * The log source identifier.
-     */
-    source: string;
-    /**
-     * The maximum number of messages to store.
-     */
-    maxLength: number;
-  }
-}
-
-/**
- * A concrete implementation of ILoggerRegistry.
- */
-export class LoggerRegistry implements ILoggerRegistry {
-  /**
-   * Construct a LoggerRegistry.
-   *
-   * @param defaultRendermime - Default rendermime to render outputs
-   * with when logger is not supplied with one.
-   */
-  constructor(options: LoggerRegistry.IOptions) {
-    this._defaultRendermime = options.defaultRendermime;
-    this._maxLength = options.maxLength;
-  }
-
-  /**
-   * Get the logger for the specified source.
-   *
-   * @param source - The name of the log source.
-   *
-   * @returns The logger for the specified source.
-   */
-  getLogger(source: string): ILogger {
-    const loggers = this._loggers;
-    let logger = loggers.get(source);
-    if (logger) {
-      return logger;
-    }
-
-    logger = new Logger({ source, maxLength: this.maxLength });
-    logger.rendermime = this._defaultRendermime;
-    loggers.set(source, logger);
-
-    this._registryChanged.emit('append');
-
-    return logger;
-  }
-
-  /**
-   * Get all loggers registered.
-   *
-   * @returns The array containing all registered loggers.
-   */
-  getLoggers(): ILogger[] {
-    return Array.from(this._loggers.values());
-  }
-
-  /**
-   * A signal emitted when the logger registry changes.
-   */
-  get registryChanged(): ISignal<this, ILoggerRegistryChange> {
-    return this._registryChanged;
-  }
-
-  /**
-   * The max length for loggers.
-   */
-  get maxLength(): number {
-    return this._maxLength;
-  }
-  set maxLength(value: number) {
-    this._maxLength = value;
-    this._loggers.forEach(logger => {
-      logger.maxLength = value;
-    });
-  }
-
-  private _maxLength: number;
-
-  private _loggers = new Map<string, Logger>();
-  private _registryChanged = new Signal<this, ILoggerRegistryChange>(this);
-  private _defaultRendermime: IRenderMimeRegistry = null;
-}
-
-export namespace LoggerRegistry {
-  export interface IOptions {
-    defaultRendermime: IRenderMimeRegistry;
-    maxLength: number;
-  }
-}
 
 /**
  * Log console output prompt implementation
@@ -372,61 +83,6 @@ class LogConsoleOutputArea extends OutputArea {
    * Output area model used by the widget.
    */
   readonly model: LoggerOutputAreaModel;
-}
-
-/**
- * Output Area Model implementation which is able to
- * limit number of outputs stored.
- */
-class LoggerOutputAreaModel extends OutputAreaModel {
-  constructor({ maxLength, ...options }: LoggerOutputAreaModel.IOptions) {
-    super(options);
-    this.maxLength = maxLength;
-  }
-
-  /**
-   * Add an output, which may be combined with previous output.
-   *
-   * @returns The total number of outputs.
-   *
-   * #### Notes
-   * The output bundle is copied. Contiguous stream outputs of the same `name`
-   * are combined. The oldest outputs are possibly removed to ensure the total
-   * number of outputs is at most `.maxLength`.
-   */
-  add(output: nbformat.IOutput): number {
-    super.add(output);
-    this._applyMaxLength();
-    return this.length;
-  }
-
-  /**
-   * Maximum number of outputs to store in the model.
-   */
-  set maxLength(value: number) {
-    this._maxLength = value;
-    this._applyMaxLength();
-  }
-
-  /**
-   * Manually apply length limit.
-   */
-  private _applyMaxLength() {
-    if (this.list.length > this._maxLength) {
-      this.list.removeRange(0, this.list.length - this._maxLength);
-    }
-  }
-
-  private _maxLength: number;
-}
-
-export namespace LoggerOutputAreaModel {
-  export interface IOptions extends IOutputAreaModel.IOptions {
-    /**
-     * The maximum number of messages stored.
-     */
-    maxLength: number;
-  }
 }
 
 /**
